rename fixed point
